LayoutTests/media/modern-media-controls/start-button/start-button.html

<link rel="stylesheet" href="../../../../Source/WebCore/Modules/modern-media-controls/controls/button.css" type="text/css" media="screen">
<link rel="stylesheet" href="../../../../Source/WebCore/Modules/modern-media-controls/controls/icon-button.css" type="text/css" media="screen">
<script src="../../../resources/js-test-pre.js"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/scheduler.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/layout-node.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/layout-item.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/icon-service.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/icon-button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/start-button.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

description("Testing the <code>StartButton</code> class.");

iconService.directoryPath = "../../../../Source/WebCore/Modules/modern-media-controls/images";

const startButton = new StartButton({ layoutTraits: LayoutTraits.macOS });

shouldBeEqualToString("startButton.element.localName", "button");
shouldBeTrue('startButton.element.classList.contains("start")');
shouldBe("startButton.element.childElementCount", "2");
shouldBeEqualToString("startButton.element.firstElementChild.localName", "div");
shouldBeEqualToString("startButton.element.firstElementChild.className", "background");
shouldBeEqualToString("startButton.element.lastElementChild.localName", "img");
shouldBeTrue("startButton.element.lastElementChild.src.includes('macOS/start@')");

const iOSStartButton = new StartButton({ layoutTraits: LayoutTraits.iOS });
shouldBeTrue("iOSStartButton.element.lastElementChild.src.includes('iOS/start@')");


</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
