LayoutTests/media/modern-media-controls/start-button/start-button-expected.txt

Testing the StartButton class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS startButton.element.localName is "button"
PASS startButton.element.classList.contains("start") is true
PASS startButton.element.childElementCount is 2
PASS startButton.element.firstElementChild.localName is "div"
PASS startButton.element.firstElementChild.className is "background"
PASS startButton.element.lastElementChild.localName is "img"
PASS startButton.element.lastElementChild.src.includes('macOS/start@') is true
PASS iOSStartButton.element.lastElementChild.src.includes('iOS/start@') is true
PASS startButton.element.getAttribute('aria-label') is "Start"
PASS successfullyParsed is true

TEST COMPLETE
